Added Research Project 26/08/25
Include the research project for our prelim and some other practices

+++ Basic Calculator/src/Main/Subtraction.java
@@ -1,8 +1,8 @@
 package Main;
 
 public class Subtraction {
-	public double subtractNumbers(double numDouble1, double numDouble2) {
-		double sum;
+	public int subtractNumbers(int numDouble1, int numDouble2) {
+		int sum;
 		sum = numDouble1 - numDouble2;
 		return sum;
 	}
